move request object to a separate file
